fixed found by live integration
To me it looks like mockgoose can be subtly different from real mongo, possibly due to version differences since production database on compose.io is not the latest. Whatever the cause, once I used this library in our production I ran into issues, so these changes are based on actual production requirements. It sucks that I can't find a way to test the changes, but e.g. mockgoose simply returns a proper promise on obj.update(), but that didn't seem to be the case in prod.. so what's a guy to do but make the necessary changes and commit?